refactor: BottleWebViewType web view feature로 이동 및 토큰 파라미터 추가
